requirements.txt
absl-py==2.3.1
aiohappyeyeballs==2.6.1
aiohttp==3.12.14
aiosignal==1.4.0
altair==5.5.0
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
bert-score==0.3.13
blinker==1.9.0
cachetools==6.1.0
certifi==2025.7.14
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
contourpy==1.3.3
cycler==0.12.1
dataclasses-json==0.6.7
datasets==4.0.0
dill==0.3.8
distro==1.9.0
evaluate==0.4.5
faiss-cpu==1.11.0
fastapi==0.116.1
filelock==3.18.0
fonttools==4.59.0
frozenlist==1.7.0
fsspec==2025.3.0
gitdb==4.0.12
GitPython==3.1.45
greenlet==3.2.3
groq==0.30.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.1
huggingface-hub==0.34.1
idna==3.10
Jinja2==3.1.6
joblib==1.5.1
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
kiwisolver==1.4.8
langchain==0.3.26
langchain-community==0.3.27
langchain-core==0.3.72
langchain-groq==0.3.6
langchain-huggingface==0.3.0
langchain-text-splitters==0.3.9
langgraph==0.5.4
langgraph-checkpoint==2.1.1
langgraph-prebuilt==0.5.2
langgraph-sdk==0.1.74
langsmith==0.4.8
MarkupSafe==3.0.2
marshmallow==3.26.1
matplotlib==3.10.3
mpmath==1.3.0
multidict==6.6.3
multiprocess==0.70.16
mypy_extensions==1.1.0
narwhals==2.0.0
networkx==3.5
nltk==3.9.1
numpy==2.3.2
orjson==3.11.1
ormsgpack==1.10.0
packaging==25.0
pandas==2.3.1
pillow==11.3.0
propcache==0.3.2
protobuf==6.31.1
pyarrow==21.0.0
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
pydeck==0.9.1
pyparsing==3.2.3
pypdf==5.9.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
PyYAML==6.0.2
referencing==0.36.2
regex==2024.11.6
requests==2.32.4
requests-toolbelt==1.0.0
rouge_score==0.1.2
rpds-py==0.26.0
safetensors==0.5.3
scikit-learn==1.7.1
scipy==1.16.1
sentence-transformers==5.0.0
setuptools==80.9.0
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
SQLAlchemy==2.0.41
starlette==0.47.2
streamlit==1.47.1
sympy==1.14.0
tenacity==9.1.2
threadpoolctl==3.6.0
tokenizers==0.21.2
toml==0.10.2
torch==2.7.1
tornado==6.5.1
tqdm==4.67.1
transformers==4.54.0
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.35.0
watchdog==6.0.0
xxhash==3.5.0
yarl==1.20.1
zstandard==0.23.0
